Add simple ui
Added a main wrapper with ".wrapper" class. Using javascript to add elements
to the DOM. Added buttons for player's selection. And added classes to
relevant elements.

Removed the loop which runs 5 rounds through loop as we'll be playing
by GUI and game will end when one user reaches a score of 5.